requirements.txt
browser-use
openai
python-dotenv
